# improve method of StackingVertically performance