--- a/tests/OpenVINO.Net.Benchmark/OpenVINO.Net.Benchmark.csproj
+++ b/tests/OpenVINO.Net.Benchmark/OpenVINO.Net.Benchmark.csproj
@@ -1,0 +1,17 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>net7.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="BenchmarkDotNet" Version="0.13.10" />
+    <PackageReference Include="BenchmarkDotNet.Annotations" Version="0.13.10" />
+    <PackageReference Include="OpenCvSharp4" Version="4.8.0.20230708" />
+    <PackageReference Include="OpenCvSharp4.slim.runtime.windows-x64" Version="4.8.0.20230708" />
+  </ItemGroup>  
+
+</Project>