--- a/tests/OpenVINO.Net.Benchmark/Program.cs.cs
+++ b/tests/OpenVINO.Net.Benchmark/Program.cs.cs
@@ -1,0 +1,10 @@
+﻿namespace OpenVINO.Net.Benchmark
+{
+    internal class Program
+    {
+        public static void Main(string[] args)
+        {
+            Console.WriteLine("Hello World!");
+        }
+    }
+}