… tests/OpenVINO.Net.Benchmark/images/test_1.jpg
… tests/OpenVINO.Net.Benchmark/images/test_2.jpg
… tests/OpenVINO.Net.Benchmark/images/test_3.jpg
… tests/OpenVINO.Net.Benchmark/images/test_4.jpg